+++ lib/michelson_parser/parser/nearley.dart
@@ -1,0 +1,550 @@
+import 'dart:convert';
+import 'dart:math';
+
+import 'package:tezster_dart/michelson_parser/grammar/michelson_grammar_tokenizer.dart';
+
+var fail = {};
+
+class Nearley {
+  NearleyGrammar grammar;
+  var options;
+
+  var lexer;
+  var lexerState;
+  var table;
+
+  int current;
+  var results;
+
+  static NearleyGrammar fromCompiled(Map<String, dynamic> rules, {start}) {
+    var lexer = rules['Lexer'];
+    start = rules['ParserStart'];
+    var redefinedRules = rules['ParserRules'].toList();
+    var _rules = redefinedRules.map<Rule>((r) {
+      return (new Rule(r['name'], r['symbols'], r['postprocess']));
+    }).toList();
+    var g = NearleyGrammar(_rules, start);
+    g.lexer = lexer;
+    return g;
+  }
+
+  void parser(NearleyGrammar rules, {start, options}) {
+    grammar = rules;
+
+    options = {
+      'keepHistory': false,
+      'lexer': grammar.lexer ?? new StreamLexer(),
+    };
+
+    lexer = options['lexer'];
+    lexerState = null;
+
+    var column = new Column(grammar, 0);
+    column.wants[grammar.start] = [];
+    column.predict(grammar.start);
+    column.process();
+    table = [column];
+    current = 0;
+  }
+
+  feed(chunk) {
+    var lexer = this.lexer;
+    lexer.tokens = <GrammarResultModel>[];
+    lexer.numindex = 0;
+    lexer.reset(chunk);
+    var token = lexer.next();
+    var column;
+    while (token != null) {
+      column = table[current];
+
+      if (current != 0) table[current - 1] = null;
+
+      var n = current + 1;
+      var nextColumn = new Column(grammar, n);
+
+      var literal = token['text'] != null ? token['text'] : token['value'];
+      var value = lexer is StreamLexer ? token['value'] : token;
+      var scannable = column.scannable;
+      for (var w = scannable.length - 1; 0 <= w; w--) {
+        var state = scannable[w];
+        var expect = state.rule.symbols[state.dot];
+        // Try to consume the token
+        // either regex or literal
+        if (expect is RegExp
+            ? expect.hasMatch(value['value'])
+            : expect['type'] != null
+                ? expect['type'] == token['type']
+                : expect['literal'] == literal) {
+          // Add it
+          var next = state.nextState({
+            'data': value,
+            'token': token,
+            'isToken': true,
+            'reference': n - 1
+          });
+          nextColumn.states.add(next);
+        }
+      }
+
+      // Next, for each of the rules, we either
+      // (a) complete it, and try to see if the reference row expected that
+      //     rule
+      // (b) predict the next nonterminal it expects by adding that
+      //     nonterminal's start state
+      // To prevent duplication, we also keep track of rules we have already
+      // added
+
+      nextColumn.process();
+      table.add(nextColumn);
+
+      // If needed, throw an error:
+      if (nextColumn.states.length == 0) {
+        var err = new NearleyError(reportError(token));
+        err.offset = this.current;
+        err.token = token;
+        throw Exception(err.error);
+      }
+
+      token = lexer.next();
+      current++;
+    }
+
+    if (column != null) {
+      this.lexerState = lexer.save();
+    }
+
+    // Incrementally keep track of results
+    results = this.finish();
+
+    return this;
+  }
+
+  finish() {
+    // Return the possible parsings
+    var considerations = [];
+    var start = this.grammar.start;
+    var column = this.table[this.table.length - 1];
+    column.states.forEach((t) {
+      if (t.rule.name == start &&
+          t.dot == t.rule.symbols.length &&
+          t.reference == 0 &&
+          t.data != fail) {
+        considerations.add(t);
+      }
+    });
+    return considerations.map((c) {
+      return c.data;
+    }).toList();
+  }
+
+  reportError(token) {
+    // var lexerMessage = this.lexer.formatError(token, "Syntax error");
+    var lines = [];
+    var tokenDisplay =
+        (token['type'] != null ? token['type'] + " token: " : "") +
+            jsonEncode(token['value'] != null ? token['value'] : token);
+    lines.add(this.lexer.formatError(token, "Syntax error"));
+    lines.add('Unexpected ' +
+        tokenDisplay +
+        '. Instead, I was expecting to see one of the following:\n');
+    var lastColumnIndex = this.table.length - 2;
+    var lastColumn = this.table[lastColumnIndex];
+    var expectantStates = lastColumn.states.where((state) {
+      if (state.rule.symbols.isNotEmpty) {
+        var nextSymbol =
+            state.rule.symbols[state.dot - 1 < 0 ? 0 : state.dot - 1];
+        return nextSymbol != null && !(nextSymbol is String);
+      } else {
+        return false;
+      }
+    }).toList();
+
+    // Display a "state stack" for each expectant state
+    // - which shows you how this state came to be, step by step.
+    // If there is more than one derivation, we only display the first one.
+    var stateStacks = expectantStates.map((state) {
+      return this.buildFirstStateStack(state, []);
+    }).toList();
+    // Display each state that is expecting a terminal symbol next.
+    stateStacks.forEach((stateStack) {
+      var state = stateStack[0];
+      var nextSymbol =
+          state.rule.symbols[state.dot - 1 < 0 ? 0 : state.dot - 1];
+      var symbolDisplay = this.getSymbolDisplay(nextSymbol);
+      lines.add('A ' + symbolDisplay + ' based on:');
+      this.displayStateStack(stateStack, lines);
+    });
+
+    lines.add("");
+    return lines.join("\n");
+  }
+
+  String reportLexerError(lexerError) {
+    var tokenDisplay, lexerMessage;
+    // Planning to add a token property to moo's thrown error
+    // even on erroring tokens to be used in error display below
+    var token = lexerError.token;
+    if (token) {
+      tokenDisplay = "input " + jsonEncode(token.text[0]) + " (lexer error)";
+      lexerMessage = this.lexer.formatError(token, "Syntax error");
+    } else {
+      tokenDisplay = "input (lexer error)";
+      lexerMessage = lexerError.message;
+    }
+    return this.reportErrorCommon(lexerMessage, tokenDisplay);
+  }
+
+  reportErrorCommon(lexerMessage, tokenDisplay) {
+    var lines = [];
+    lines.add(lexerMessage);
+    var lastColumnIndex = this.table.length - 1;
+    var lastColumn = this.table[lastColumnIndex];
+    var expectantStates = lastColumn.states.where((state) {
+      var nextSymbol = state.rule.symbols[state.dot] ?? null;
+      return nextSymbol != null && !(nextSymbol is String);
+    }).toList();
+
+    if (expectantStates.length == 0) {
+      lines.add('Unexpected ' +
+          tokenDisplay +
+          '. I did not expect any more input. Here is the state of my parse table:\n');
+      this.displayStateStack(lastColumn.states, lines);
+    } else {
+      lines.add('Unexpected ' +
+          tokenDisplay +
+          '. Instead, I was expecting to see one of the following:\n');
+      // Display a "state stack" for each expectant state
+      // - which shows you how this state came to be, step by step.
+      // If there is more than one derivation, we only display the first one.
+      var stateStacks = expectantStates.map((state) {
+        return this.buildFirstStateStack(state, []) ?? [state];
+      });
+      // Display each state that is expecting a terminal symbol next.
+      stateStacks.forEach((stateStack) {
+        var state = stateStack[0];
+        var nextSymbol = state.rule.symbols[state.dot];
+        var symbolDisplay = this.getSymbolDisplay(nextSymbol);
+        lines.add('A ' + symbolDisplay + ' based on:');
+        this.displayStateStack(stateStack, lines);
+      });
+    }
+    lines.add("");
+    return lines.join("\n");
+  }
+
+  buildFirstStateStack(state, visited) {
+    if (visited.indexOf(state) != -1) {
+      // Found cycle, return null
+      // to eliminate this path from the results, because
+      // we don't know how to display it meaningfully
+      return null;
+    }
+    if (state.wantedBy.length == 0) {
+      return [state];
+    }
+    var prevState = state.wantedBy[0];
+    var childVisited = [state]..addAll(visited);
+    var childResult = this.buildFirstStateStack(prevState, childVisited);
+    if (childResult == null) {
+      return null;
+    }
+    return [state]..addAll(childResult);
+  }
+
+  displayStateStack(stateStack, lines) {
+    var lastDisplay;
+    var sameDisplayCount = 0;
+    for (var j = 0; j < stateStack.length; j++) {
+      var state = stateStack[j];
+      var display = state.rule.toStringWithData(state.dot);
+      if (display == lastDisplay) {
+        sameDisplayCount++;
+      } else {
+        if (sameDisplayCount > 0) {
+          lines.add('    ^ ' +
+              sameDisplayCount.toString() +
+              ' more lines identical to this');
+        }
+        sameDisplayCount = 0;
+        lines.add('    ' + display);
+      }
+      lastDisplay = display;
+    }
+  }
+
+  getSymbolDisplay(symbol) {
+    return getSymbolLongDisplay(symbol);
+  }
+
+  getSymbolLongDisplay(symbol) {
+    // var type = typeof symbol;
+    if (symbol is String) {
+      return symbol;
+    } else if (symbol is Map) {
+      return jsonEncode(symbol['literal'] ?? '');
+    } else if (symbol is RegExp) {
+      return 'character matching ' + symbol.toString();
+    } else if (symbol['type'] != null) {
+      return symbol['type'] + ' token';
+    } else if (symbol['test'] ?? null != null) {
+      return 'token matching ' + symbol['test'].toString();
+    } else {
+      throw new Exception('Unknown symbol type: ' + symbol.toString());
+    }
+  }
+}
+
+class Column {
+  NearleyGrammar grammar;
+  int index;
+
+  List<ColumnState> states;
+  Map wants;
+  List<ColumnState> scannable;
+  Map completed;
+
+  Column(this.grammar, this.index) {
+    states = [];
+    wants = {};
+    scannable = [];
+    completed = {};
+  }
+
+  void predict(exp) {
+    var rules = grammar.byName[exp] ?? [];
+    for (var i = 0; i < rules.length; i++) {
+      var r = rules[i];
+      var wantBy = wants[exp];
+      var s = new ColumnState(r, 0, index, wantBy);
+      states.add(s);
+    }
+  }
+
+  void process() {
+    var _states = this.states;
+    var _wants = this.wants;
+    var _completed = this.completed;
+
+    for (var w = 0; w < _states.length; w++) {
+      var state = _states[w];
+      if (state.isComplete) {
+        state.finish();
+        if (state.data != fail) {
+          var wantBy = state.wantedBy;
+
+          // complete
+          for (var i = wantBy.length - 1; 0 <= i; i--) {
+            var left = wantBy[i];
+            this.complete(left, state);
+          }
+
+          // special-case nullables
+          if (state.reference == index) {
+            // make sure future predictors of this rule get completed.
+            var exp = state.rule.name;
+            this.completed[exp] = this.completed[exp] ?? [];
+            this.completed[exp].add(state);
+          }
+        }
+      } else {
+        // queue scannable states
+        var exp = state.rule.symbols[state.dot];
+        if (!(exp is String)) {
+          this.scannable.add(state);
+        }
+
+        // predict
+        if (_wants[exp] != null) {
+          _wants[exp].add(state);
+          if (_completed.containsKey(exp)) {
+            var nulls = _completed[exp];
+            for (var i = 0; i < nulls.length; i++) {
+              var right = nulls[i];
+              complete(state, right);
+            }
+          }
+        } else {
+          _wants[exp] = _wants[exp] is List ? _wants[exp].add(state) : [state];
+          this.predict(exp);
+        }
+      }
+    }
+  }
+
+  void complete(left, right) {
+    var copy = left.nextState(right);
+    states.add(copy);
+  }
+}
+
+class ColumnState {
+  Rule rule;
+  var dot;
+  var reference;
+  var data;
+  var wantedBy;
+  var isComplete;
+
+  ColumnState left;
+  var right;
+
+  ColumnState(this.rule, this.dot, this.reference, this.wantedBy) {
+    data = [];
+    isComplete = dot == rule.symbols.length;
+  }
+
+  void finish() {
+    if (rule.postprocess != null) {
+      try {
+        data = rule.postprocess(data);
+      } catch (e) {
+        print("Error In ===> " + rule.name);
+      }
+    }
+  }
+
+  ColumnState nextState(child) {
+    var state = new ColumnState(rule, dot + 1, reference, wantedBy);
+    state.left = this;
+    state.right = child;
+    if (state.isComplete) {
+      state.data = state.build();
+    }
+    return state;
+  }
+
+  List build() {
+    var children = [];
+    var node = this;
+    do {
+      children.add(
+          node.right is ColumnState ? node.right.data : node.right['data']);
+      node = node.left;
+    } while (node.left != null);
+    children = children.reversed.toList();
+    return children;
+  }
+}
+
+class NearleyError {
+  var error;
+  int offset;
+  var token;
+  NearleyError(this.error);
+}
+
+class NearleyGrammar {
+  List<Rule> rules;
+  var start;
+
+  var lexer;
+  var byName;
+
+  NearleyGrammar(this.rules, this.start) {
+    this.rules = rules;
+    this.start = start ?? this.rules[0].name;
+    var byName = this.byName = {};
+    this.rules.forEach((rule) {
+      if (!byName.containsKey(rule.name)) {
+        byName[rule.name] = [];
+      }
+      byName[rule.name].add(rule);
+    });
+  }
+}
+
+class Rule {
+  var name;
+  var symbols;
+  var postprocess;
+  Rule(this.name, this.symbols, this.postprocess);
+
+  toStringWithData(withCursorAt) {
+    stringifySymbolSequence(e) {
+      return e is Map && e.containsKey('literal')
+          ? jsonEncode(e['literal'])
+          : e is Map && e.containsKey('type')
+              ? '%' + e['type']
+              : e.toString();
+    }
+
+    var symbolSequence = (withCursorAt == null)
+        ? this.symbols.map(stringifySymbolSequence).join(' ')
+        : (this
+                .symbols
+                .take(withCursorAt - 1 < 0 ? 0 : withCursorAt - 1)
+                .map(stringifySymbolSequence)
+                .join(' ') +
+            " ● " +
+            this
+                .symbols
+                .take(withCursorAt)
+                .map(stringifySymbolSequence)
+                .join(' '));
+    return this.name + " → " + symbolSequence;
+  }
+}
+
+class StreamLexer {
+  var buffer;
+  int index;
+  var line;
+  var lastLineBreak;
+
+  StreamLexer() {
+    this.reset("");
+  }
+
+  has(tokenType) {
+    return true;
+  }
+
+  reset(data, {state}) {
+    this.buffer = data;
+    this.index = 0;
+    this.line = state != null ? state.line : 1;
+    this.lastLineBreak = state != null ? -state.col : 0;
+  }
+
+  save() {
+    return {
+      'line': this.line,
+      'col': this.index - this.lastLineBreak,
+    };
+  }
+
+  formatError(token, message) {
+    pad(n, length) {
+      var s = (n).toString();
+      return List.generate(length - s.length + 1, (index) => '').join(" ") + s;
+    }
+
+    // nb. this gets called after consuming the offending token,
+    // so the culprit is index-1
+    var buffer = this.buffer;
+    if (buffer is String) {
+      var lines = buffer.split("\n").sublist(max(0, this.line - 5), this.line);
+
+      var nextLineBreak = buffer.indexOf('\n', this.index);
+      if (nextLineBreak == -1) nextLineBreak = buffer.length;
+      var col = this.index - this.lastLineBreak;
+      var lastLineDigits = (this.line).toString().length;
+      message += " at line " +
+          this.line.toString() +
+          " col " +
+          col.toString() +
+          ":\n\n";
+      var msg = List();
+      for (var i = 0; i < lines.length; i++) {
+        var line = lines[i];
+        msg.add(
+            pad(this.line - lines.length + i + 1, lastLineDigits) + " " + line);
+      }
+      message += msg.join("\n");
+      message += "\n" + pad("", lastLineDigits + col) + "^\n";
+      return message;
+    } else {
+      return message + " at index " + (this.index - 1);
+    }
+  }
+}