+++ test/tezos_test.dart
@@ -51,7 +51,7 @@ void main() {
       publicKeyHash: 'tz1LRibbLEEWpaXb4aKrXXgWPvx9ue9haAAV',
     );
 
-    var signer = await TezsterDart.createSigner(
+    await TezsterDart.createSigner(
         TezsterDart.writeKeyWithHint(keyStore.secretKey, 'edsk'));
   });